Add ajax handlers and edit view for ads

+++ app/public/wp-content/plugins/tta-management-plugin/includes/admin/views/ads-create.php
@@ -15,27 +15,36 @@ if ( isset( $_GET['ad_id'] ) && isset( $ads[ intval( $_GET['ad_id'] ) ] ) ) {
 }
 
 if ( isset( $_POST['tta_ad_save'] ) && check_admin_referer( 'tta_ad_save_action', 'tta_ad_save_nonce' ) ) {
-    $new_ad = [
-        'image_id'        => intval( $_POST['image_id'] ?? 0 ),
-        'url'             => esc_url_raw( $_POST['url'] ?? '' ),
-        'business_name'   => sanitize_text_field( $_POST['business_name'] ?? '' ),
-        'business_phone'  => sanitize_text_field( $_POST['business_phone'] ?? '' ),
-        'business_address'=> sanitize_text_field( $_POST['business_address'] ?? '' ),
-    ];
-
-    if ( $editing && null !== $index ) {
-        $ads[ $index ] = $new_ad;
-        echo '<div class="updated"><p>' . esc_html__( 'Ad updated.', 'tta' ) . '</p></div>';
-    } else {
-        $ads[] = $new_ad;
-        $index = array_key_last( $ads );
-        $editing = true;
-        echo '<div class="updated"><p>' . esc_html__( 'Ad created.', 'tta' ) . '</p></div>';
+    $url = sanitize_text_field( $_POST['url'] ?? '' );
+    if ( $url && ! preg_match( '#^https?://#i', $url ) ) {
+        $url = 'https://' . $url;
     }
+    $host = parse_url( $url, PHP_URL_HOST );
+    if ( ! $host || ! preg_match( '/\.[a-z]{2,}$/i', $host ) ) {
+        echo '<div class="error"><p>' . esc_html__( 'Please enter a valid Link URL.', 'tta' ) . '</p></div>';
+    } else {
+        $new_ad = [
+            'image_id'        => intval( $_POST['image_id'] ?? 0 ),
+            'url'             => esc_url_raw( $url ),
+            'business_name'   => sanitize_text_field( $_POST['business_name'] ?? '' ),
+            'business_phone'  => sanitize_text_field( $_POST['business_phone'] ?? '' ),
+            'business_address'=> sanitize_text_field( $_POST['business_address'] ?? '' ),
+        ];
 
-    update_option( 'tta_ads', array_values( $ads ), false );
-    TTA_Cache::delete( 'tta_ads_all' );
-    $ad = $new_ad;
+        if ( $editing && null !== $index ) {
+            $ads[ $index ] = $new_ad;
+            echo '<div class="updated"><p>' . esc_html__( 'Ad updated.', 'tta' ) . '</p></div>';
+        } else {
+            $ads[] = $new_ad;
+            $index   = array_key_last( $ads );
+            $editing = true;
+            echo '<div class="updated"><p>' . esc_html__( 'Ad created.', 'tta' ) . '</p></div>';
+        }
+
+        update_option( 'tta_ads', array_values( $ads ), false );
+        TTA_Cache::delete( 'tta_ads_all' );
+        $ad = $new_ad;
+    }
 }
 ?>
 <div class="wrap">
@@ -49,14 +58,14 @@ if ( isset( $_POST['tta_ad_save'] ) && check_admin_referer( 'tta_ad_save_action'
             <tr>
                 <th scope="row"><?php esc_html_e( 'Ad Image', 'tta' ); ?></th>
                 <td>
-                    <button class="button tta-upload-single" data-target="#ad_image"><?php esc_html_e( 'Select Image', 'tta' ); ?></button>
-                    <input type="hidden" id="ad_image" name="image_id" value="<?php echo esc_attr( $ad['image_id'] ?? 0 ); ?>">
-                    <div id="ad_image_preview"><?php echo ! empty( $ad['image_id'] ) ? wp_get_attachment_image( intval( $ad['image_id'] ), 'thumbnail' ) : ''; ?></div>
+                    <button class="button tta-upload-single" data-target="#ad_imageid"><?php esc_html_e( 'Select Image', 'tta' ); ?></button>
+                    <input type="hidden" id="ad_imageid" name="image_id" value="<?php echo esc_attr( $ad['image_id'] ?? 0 ); ?>">
+                    <div id="ad_image-preview"><?php echo ! empty( $ad['image_id'] ) ? wp_get_attachment_image( intval( $ad['image_id'] ), 'thumbnail' ) : ''; ?></div>
                 </td>
             </tr>
             <tr>
                 <th scope="row"><label for="url"><?php esc_html_e( 'Link URL', 'tta' ); ?></label></th>
-                <td><input type="url" name="url" id="url" class="regular-text" value="<?php echo esc_attr( $ad['url'] ?? '' ); ?>"></td>
+                <td><input type="text" name="url" id="url" class="regular-text" value="<?php echo esc_attr( $ad['url'] ?? '' ); ?>" placeholder="example.com"></td>
             </tr>
             <tr>
                 <th scope="row"><label for="business_name"><?php esc_html_e( 'Business Name', 'tta' ); ?></label></th>
